# [READY] perf: view as named tuples