Makes `View` into `NamedTuple` s

Part 1 of https://github.com/geohot/tinygrad/pull/1044 that does not require to modify all the tests using ShapeTracker.

Pure refactor.