@roelofvandijk 2023-07-05
This branch
```
codegen         mean runtime:  110.57ms, runs:   131.90,  108.14,  105.87,  135.21,  102.50,  101.04,  101.29,  103.84,  104.31,  111.59
methodcache     mean runtime:  102.97ms, runs:    99.29,  104.39,   97.48,  133.32,   98.55,   97.38,   96.08,   97.32,   98.74,  107.12
profile         mean runtime:  441.84ms, runs:   464.32,  425.24,  435.10,  430.19,  435.78,  426.87,  434.37,  475.61,  428.63,  462.32

574.05 lazy.py
556.54 tensor.py
108.67 shapetracker.py
13.86 ops.py
13.73 symbolic.py
244.78 helpers.py
8.58 lib.py
55.33 linearizer.py
1575.53 total
```


Master (981d4980c4e8e23b54a6a2e4b4d6cfc8c12adea7)
```
codegen         mean runtime:  122.19ms, runs:   131.45,  109.08,  143.21,  111.72,  114.41,  117.49,  114.22,  156.37,  113.04,  110.86
methodcache     mean runtime:  121.16ms, runs:   107.58,  109.79,  117.69,  163.99,  114.72,  110.76,  106.83,  104.62,  117.10,  158.53
profile         mean runtime:  510.99ms, runs:   605.95,  515.10,  510.84,  501.23,  512.01,  523.03,  460.30,  492.48,  481.04,  507.88

617.03 lazy.py
598.30 tensor.py
156.17 shapetracker.py
15.95 ops.py
44.49 symbolic.py
244.51 helpers.py
8.85 lib.py
60.72 linearizer.py
1746.00 total
```