@geohot 2023-07-07
This is looking much better :) Let me read it