@roelofvandijk 2023-07-08
New timings

```
Master
codegen         mean runtime:  126.11ms, runs:   134.24,  119.01,  167.93,  124.40,  116.67,  112.67,  114.14,  151.54,  111.23,  109.22
methodcache     mean runtime:  113.07ms, runs:   108.58,  108.07,  108.02,  146.36,  105.10,  100.91,  102.42,  104.86,  104.47,  141.90
profile         mean runtime:  519.42ms, runs:   559.09,  513.73,  509.23,  503.66,  527.16,  560.75,  514.92,  482.66,  525.01,  497.97

As module functions
codegen         mean runtime:  111.63ms, runs:   126.70,  106.61,  103.87,  135.38,  104.26,  105.57,  106.09,  107.41,  111.63,  108.85
methodcache     mean runtime:  106.92ms, runs:   102.26,  102.26,   99.32,  136.89,   99.58,   97.80,  102.90,  116.02,  108.47,  103.69
profile         mean runtime:  467.54ms, runs:   532.60,  477.76,  440.87,  466.60,  480.84,  460.97,  439.55,  464.00,  445.49,  466.70

As staticmethods
codegen         mean runtime:  111.59ms, runs:   132.02,  106.09,  104.19,  145.98,  104.14,  102.51,  103.49,  104.46,  108.30,  104.68
methodcache     mean runtime:  106.52ms, runs:   114.48,  110.00,  108.61,  140.94,   97.78,   95.75,   97.81,   99.53,  100.43,   99.90
profile         mean runtime:  458.36ms, runs:   484.93,  446.32,  448.09,  458.38,  431.30,  444.64,  436.47,  500.83,  457.25,  475.37

```