@roelofvandijk 2023-07-08
Because I had removed several methods that were used only once, mainly in the View init.
* `filter_strides` 
* `is_contiguous`
* `view_from_shape`
Doing so reduces the function call overhead.

I have stripped down everything to the bare necessities - should be a tidy diff now. 

All other optimizations will come in a separate MR.