@roelofvandijk 2023-07-08
```
LAZYCACHE=0 python3.11 -O test/external/external_test_speed_llama.py


Master
codegen         mean runtime:  119.23ms, runs:   135.19,  111.06,  138.51,  110.68,  108.25,  112.41,  111.19,  143.45,  109.42,  112.13
methodcache     mean runtime:  111.93ms, runs:   105.79,  105.06,  105.42,  140.35,  104.80,  102.98,  104.12,  104.21,  102.70,  143.83
profile         mean runtime:  490.74ms, runs:   554.70,  459.67,  507.87,  465.26,  460.97,  503.71,  463.13,  467.14,  507.11,  517.83

Thise branch: bare necessities (view as tuples, staticmethod)
codegen         mean runtime:  110.95ms, runs:   127.21,  103.91,  102.17,  134.03,  106.16,  106.11,  102.39,  102.14,  113.83,  111.61
methodcache     mean runtime:  101.65ms, runs:    98.15,  102.76,  100.76,  133.84,   97.82,   96.36,   95.35,   94.60,   99.47,   97.37
profile         mean runtime:  477.18ms, runs:   472.95,  433.84,  498.87,  467.14,  467.92,  485.08,  476.20,  532.79,  474.53,  462.44
```