@roelofvandijk 2023-07-09
@geohot This should be done.

I think I can achieve a 20% speedup on LLaMA (and basically everything else) with this and similar refactors (e.g. remove the ShapeTracker class, better caching, faster Function/Tensor/Buffer initialization). 

Does that sound worthy of a bounty to you? E.g. <80ms on LLaMa?