@geohot 2023-07-09
Like I sort of feel this speed isn't worth it. I'm fine with NamedTuple for the immutability, but the last PR probably went a bit too far.