@roelofvandijk 2023-07-09
Re: speed - I understand the sentiment, because it's only 3% here, 5% there. 

It adds up though. I'll make an MR to show how much fruit is left.