@geohot 2023-07-09
Hmm replacing ShapeTracker with a tuple of views is interesting...