@geohot 2023-07-09
Bounty added. https://github.com/tinygrad/tinygrad/issues/1209